Test env

+++ CurriculumManagementSystemWebAPI/Controllers/AssessmentMethodsController.cs
@@ -31,6 +31,14 @@ namespace CurriculumManagementSystemWebAPI.Controllers
             return Ok(new BaseResponse(false, "List Assessment Method", listAssessmentMethodResponse));
         }
 
+        [HttpGet("GetAssessmentMethodById/{id}")]
+        public ActionResult GetAssessmentMethod(int id)
+        {
+            var assessmentMethod = assessmentMethodRepository.GetAsssentMethodById(id);
+            var assessmentMethodResponse = _mapper.Map<AssessmentMethodDTOResponse>(assessmentMethod);
+            return Ok(new BaseResponse(false, "Assessment Method", assessmentMethodResponse));
+        }
+
 
         [HttpGet("Pagination/{page}/{limit}")]
         public ActionResult PaginationAssessmentMethod(int page, int limit, [FromQuery] string? txtSearch)
@@ -40,14 +48,15 @@ namespace CurriculumManagementSystemWebAPI.Controllers
             {
                 Ok(new BaseResponse(false, "Not Found Assessment Method!"));
             }
+            var total = assessmentMethodRepository.GetTotalAssessmentMethod(txtSearch);
             var listAssessmentMethodResponse = _mapper.Map<List<AssessmentMethodDTOResponse>>(listAssessmentMethod);
-            return Ok(new BaseResponse(false, "List Assessment Method", listAssessmentMethodResponse));
+            return Ok(new BaseResponse(false, "List Assessment Method", new BaseListResponse(page, limit, total, listAssessmentMethodResponse)));
         }
 
         [HttpPost("CreateAssessmentMethod")]
         public ActionResult CreateAssessmentMethod([FromBody] AssessmentMethodRequest assessmentMethodRequest)
         {
-            if (assessmentMethodRepository.CheckAssmentMethodDuplicate(assessmentMethodRequest.assessment_method_component))
+            if (assessmentMethodRepository.CheckAssmentMethodDuplicate(0,assessmentMethodRequest.assessment_method_component))
             {
                 return BadRequest(new BaseResponse(true, "Assessment Method Duplicate!"));
             }
@@ -75,7 +84,7 @@ namespace CurriculumManagementSystemWebAPI.Controllers
             }
 
 
-            if (assessmentMethodRepository.CheckAssmentMethodDuplicate(assessmentMethodRequest.assessment_method_component))
+            if (assessmentMethodRepository.CheckAssmentMethodDuplicate(id, assessmentMethodRequest.assessment_method_component))
             {
                 return BadRequest(new BaseResponse(true, "Assessment Method Duplicate!"));
             }
@@ -104,11 +113,18 @@ namespace CurriculumManagementSystemWebAPI.Controllers
                 return NotFound(new BaseResponse(true, "Can't Found Assessment Method"));
             }
 
-            string createResult = assessmentMethodRepository.DeleteAssessmentMethod(assessmentMethod);
 
-            if (!createResult.Equals(Result.deleteSuccessfull.ToString()))
+            if (assessmentMethodRepository.CheckAssmentMethodExsit(id))
             {
-                return BadRequest(new BaseResponse(true, createResult));
+                return BadRequest(new BaseResponse(true, "Assessment Method is Used! Can't Delete"));
+            }
+
+
+            string deleteResult = assessmentMethodRepository.DeleteAssessmentMethod(assessmentMethod);
+
+            if (!deleteResult.Equals(Result.deleteSuccessfull.ToString()))
+            {
+                return BadRequest(new BaseResponse(true, deleteResult));
             }
 
             var assessMethodRespone = _mapper.Map<AssessmentMethodDTOResponse>(assessmentMethod);
